v1.6.1 - Custom-ish music
# Changes

Changed Record 13's music:

C418 - 13 -> C418 - 13 / Excuse

Changed Record Cat's music:

C418 - Cat -> C418 - Cat / Dog

# Why such a change?

I wanted to keep the spirit of these two records as the way they are, but I wanted to add their transitioning hidden tracks to make them even more special, especially with 13.. not many people really enjoy 13's ambience but the addition of Excuse adds the reward of a calming yet melancholic tune.

Cat / Dog is actually in the Legacy Console Editions of Minecraft, and I just wanted to implement that too.

# Issues Stemming From Change

one notable issue I found, and patched (hence why its 1.6.ONE), was audio issues with differentiating between .ogg and .mus (.ogg is always played by jukeboxes....). The .mus was taking up the game's sound memory and when it was cached into the jukebox, it did not read it, throwing an error.

However, that is since fixed (totally didn't take ~5 hours of juggling).

The only issue I found with the records is that the sounds don't source or play from the jukebox, it basically plays the sound as if it were a background song, AKA globally.

That issue will be in the GitHub issues as OAP-2.

It will be my personal preference to patch that, since I hear some players actually prefer their jukebox sounds to be played globally, since the song from the jukebox has a small-ish range and only plays out of one ear when looking perpendicular to it.

Good Day. :)