1.7.0 - Double Build Height!!
# Changes

After a few days of juggling with the lighting engine (somehow the root of my problems, not terrain..) I was able to double the build height limit (from 128 to 256) :)

# Notice

I cannot promise this will run perfectly though.. I noticed a few chunk errors whenever I was lagging, though that could because I was emulating an x86_64 java 8 on an ARM64 computer and it was lagging behind.

Nonetheless, enjoy the extra height :))

If you plan to share any of your worlds to anyone, and your world includes builds that exceed the 128 limit--encourage them to get this mod. Otherwise, the build will be completely sliced from y-128 upward, and it'll be an awkward looking penthouse.. :^)

# For the future

There'll be more to come down the road if I don't get burned out from just how much work this stuff took.. I ruined my sleep schedule for this.

Again, I'll probably be moving on to more things after 2.0.0, unless I want to keep maintaining this as a sort of all-in-one mod for alpha :D. I mean, I don't see anybody else doing that, might be good for people wanting to get into Alpha Minecraft.

Good Day. :)